(#9387, #9224) Perms fix for puppet-dashboard
The postinst script chowns the whole directory recursively to puppet-dashboard,
including the top level /etc/puppetlabs/puppet-dashboard directory. This should
still be owned by root and was inconsistent with the ownership on EL/SLES.
Additionally, the postinst -x everything, including the top level directory,
for others. This broke the ENC abilities of dashboard during an upgrade.
This patch chowns the top level directory to root and chmods it to 755. This is
part of the fix of the ENC upgrade bug (#9387), the other part is in the
upgrade script itself.

Signed-off-by: Matthaus Litteken <matthaus@puppetlabs.com>