(#11571) Unify the settings.yml between EL and Debian
The Debian version of the settings.yml file was ending up with different
settings than those on EL.  The EL settings were more correct.  Note
that in many cases this doesn't matter since the PE installer overwrites
the settings yaml file; but out of the box the file should contain the
same values.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>