(#11699, #9626) Fix css, js file permissions in dashboard
css and js files that are generated by the dashboard (all.css, all.js) can have
interesting permissions, depending upon the umask. If the umask is 027, this
can cause rendering and functionality problems in the console. This fix adds
the www-data user to the puppet-dashboard group to help mitigate that problem.